Parse all of the index information from the Javadoc index files.
When climbing up to a parent element because a link is wrapped in a
style tag, also re-fetch the parent's tag name.

Doing so causes methods the check for the <dt> tag to succeed, and
all of the bolded entries in index-all.html are now seen.

Because of this, it is also no longer necessary to search class
files in order to find methods, fields, and constructors.